+++ modules/Findx.Module.Admin/Models/SysNoticeUserInfo.cs
@@ -32,7 +32,7 @@ namespace Findx.Module.Admin.Models
 		/// 状态（字典 0未读 1已读）
 		/// </summary>
 		[Column(Name = "status", DbType = "tinyint(4)")]
-		public sbyte Status { get; set; }
+		public int Status { get; set; }
 
 		/// <summary>
 		/// 用户id
